requirements.txt
asgiref==3.8.1
Django==5.1.5
django-filebrowser-no-grappelli==4.0.2
django-multiselectfield==0.1.13
django-tinymce==4.1.0
pillow==11.1.0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
wfastcgi==3.0.0
whitenoise==6.8.2
